--- src/main/java/com/handwoong/rainbowletter/config/security/AllowUri.java
@@ -1,5 +1,0 @@
-package com.handwoong.rainbowletter.config.security;
-
-public interface AllowUri {
-    String getUri();
-}
