+++ src/main/java/com/handwoong/rainbowletter/config/security/SecurityConfig.java
@@ -1,7 +1,14 @@
 package com.handwoong.rainbowletter.config.security;
 
+import com.handwoong.rainbowletter.config.security.handler.CustomAccessDeniedHandler;
+import com.handwoong.rainbowletter.config.security.handler.CustomAuthenticationEntryPoint;
+import com.handwoong.rainbowletter.config.security.jwt.JwtTokenAuthenticationFilter;
 import com.handwoong.rainbowletter.config.security.oauth.CustomOAuthUserService;
 import com.handwoong.rainbowletter.config.security.oauth.OAuthSuccessHandler;
+import com.handwoong.rainbowletter.config.security.uri.AccessAllowUri;
+import com.handwoong.rainbowletter.config.security.uri.AdminAllowUri;
+import com.handwoong.rainbowletter.config.security.uri.AllowUri;
+import com.handwoong.rainbowletter.config.security.uri.AnonymousAllowUri;
 import java.util.Arrays;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
