+++ src/main/java/com/handwoong/rainbowletter/config/security/uri/AllowUri.java
@@ -1,0 +1,5 @@
+package com.handwoong.rainbowletter.config.security.uri;
+
+public interface AllowUri {
+    String getUri();
+}
