+++ src/main/java/com/handwoong/rainbowletter/domain/mail/service/template/EmailVerifyTemplate.java
@@ -1,9 +1,9 @@
 package com.handwoong.rainbowletter.domain.mail.service.template;
 
 import com.handwoong.rainbowletter.config.PropertiesConfig;
-import com.handwoong.rainbowletter.config.security.GrantType;
-import com.handwoong.rainbowletter.config.security.JwtTokenProvider;
-import com.handwoong.rainbowletter.config.security.TokenResponse;
+import com.handwoong.rainbowletter.config.security.jwt.GrantType;
+import com.handwoong.rainbowletter.config.security.jwt.JwtTokenProvider;
+import com.handwoong.rainbowletter.config.security.jwt.TokenResponse;
 import com.handwoong.rainbowletter.domain.mail.dto.EmailTemplateDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
